Fix #17
Featured image now have `width`, `height` and `alt` attributes.

To update previous featured images, simply open the post and resave it.
The compoment automatically reload the image from the URL.

The `alt` attribute's value is taken from the image `title`
attribute of the API response.

The `width` attribute is always 770px, since it appears to be the
constant value used for thumbnail images.

The `height` attribute is calculated from the `widht` and `height` of
the full image, proportionnaly to the constant width value.